requirements.txt
annotated-types==0.7.0
anyio==4.8.0
backoff==2.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
langfuse==2.57.10
numpy==2.2.2
openai==1.59.7
packaging==24.2
portalocker==2.10.1
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
python-decouple==3.8
qdrant-client==1.13.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tabulate==0.9.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
